+++ tests/test_common.rs
@@ -6,7 +6,7 @@ mod tests {
   use meritrank::errors::MeritRankError;
   use meritrank::graph::{NodeId, EdgeId, Weight};
   use meritrank::random_walk::RandomWalk;
-  use meritrank::walk_storage::{decide_skip_invalidation_on_edge_addition, WalkStorage};
+  use meritrank::walk_storage::{decide_skip_invalidation, decide_skip_invalidation_on_edge_addition, decide_skip_invalidation_on_edge_deletion};
 
   use rand::rngs::StdRng;
   use rand::SeedableRng;
@@ -33,25 +33,116 @@ mod tests {
     assert_eq!(error.to_string(), "Random choice error");
   }
 
+
   #[test]
   fn test_decide_skip_invalidation() {
-    let storage = WalkStorage::new();
-    let walk = RandomWalk::from_nodes(vec![1, 2]);
-    let pos = 0;
-    let edge: EdgeId = (1, 2);
+    let walk = RandomWalk::from_nodes(vec![1, 2, 3]);
+    let edge: EdgeId = (2, 3);
     let step_recalc_probability = 0.5;
-
-    let rng_seed = 1234;
+    let rng_seed = 1342; // Set the seed for the random number generator
 
     // Create a deterministic random number generator
     let mut rng = StdRng::seed_from_u64(rng_seed);
 
-    let (_may_skip, _new_pos) = decide_skip_invalidation_on_edge_addition(
-      &walk,           // walk: This is the walk that is being invalidated
-      pos,  // pos: This is the position in the walk that is being invalidated
-      edge, // edge: This is the edge that is being added
-      step_recalc_probability, // The probability that the step will be recalculated
-      Some(&mut rng), // rng: This is the random number generator
+    // Test skipping invalidation on edge deletion
+    let (may_skip, new_pos) =
+        decide_skip_invalidation(&walk, 2, edge, 0.0, Some(&mut rng));
+    assert!(may_skip);
+    assert_eq!(new_pos, 2);
+
+    // Test skipping invalidation with step recalculation probability
+    let (may_skip, new_pos) = decide_skip_invalidation(
+      &walk,
+      1,
+      edge,
+      step_recalc_probability,
+      Some(&mut rng),
     );
+    assert!(may_skip);
+    assert_eq!(new_pos, 1);
+
+    // Test invalidation without skipping
+    let (may_skip, new_pos) = decide_skip_invalidation(
+      &walk,
+      0,
+      edge,
+      step_recalc_probability,
+      Some(&mut rng),
+    );
+    assert!(!may_skip);
+    assert_eq!(new_pos, 1);
+  }
+
+  #[test]
+  fn test_walk_storage_decide_skip_invalidation_on_edge_deletion() {
+    let walk = RandomWalk::from_nodes(vec![1, 2, 3]);
+    let edge: EdgeId = (2, 3);
+
+    // Test invalidation without skipping
+    let (may_skip, new_pos) = decide_skip_invalidation_on_edge_deletion(&walk, 0, edge);
+    assert!(!may_skip);
+    assert_eq!(new_pos, 1);
+
+    // Test invalidation with skipping
+    let (may_skip, new_pos) = decide_skip_invalidation_on_edge_deletion(&walk, 1, edge);
+    assert!(!may_skip);
+    assert_eq!(new_pos, 1);
+
+    // Test invalidation at the end of the walk
+    let (may_skip, new_pos) = decide_skip_invalidation_on_edge_deletion(&walk, 2, edge);
+    assert!(may_skip);
+    assert_eq!(new_pos, 2);
+  }
+
+  #[test]
+  fn test_decide_skip_invalidation_on_edge_addition() {
+    use rand::random;
+
+    let walk = RandomWalk::from_nodes(vec![
+      1,
+      2,
+      1,
+      3,
+    ]);
+    let edge: EdgeId = (1, 2);
+    let step_recalc_probability = 0.5;
+    let rng_seed = 1342; // Set the seed for the random number generator
+
+    // Create a deterministic random number generator
+    let mut rng = StdRng::seed_from_u64(rng_seed);
+
+    // Test invalidation without skipping
+    let (may_skip, new_pos) = decide_skip_invalidation_on_edge_addition(
+      &walk,
+      0,
+      edge,
+      step_recalc_probability,
+      Some(&mut rng),
+    );
+    assert!(!may_skip);
+    assert_eq!(new_pos, 2);
+
+    // Test invalidation with skipping
+    let (may_skip, new_pos) = decide_skip_invalidation_on_edge_addition(
+      &walk,
+      2,
+      edge,
+      step_recalc_probability,
+      Some(&mut rng),
+    );
+    assert!(may_skip);
+    assert_eq!(new_pos, 2);
+
+    // Test invalidation with step recalculation
+    let (may_skip, new_pos) = decide_skip_invalidation_on_edge_addition(
+      &walk,
+      1,
+      edge,
+      step_recalc_probability,
+      Some(&mut rng),
+    );
+    // let should_skip = random::<Weight>() > step_recalc_probability;
+    assert_eq!(may_skip, false);
+    assert_eq!(new_pos, 2);
   }
 }
